requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
attrs==21.4.0
    # via pytest
click==8.1.3
    # via code-annotations
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==6.4
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via code-annotations
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via pytest
pbr==5.9.0
    # via stevedore
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-slugify==6.1.2
    # via code-annotations
pytz==2022.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0
    # via code-annotations
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==3.5.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
